Multi - use
The main file has been changed: I figured out how to turn on the bot itself and leave the endless cycle of sending messages working, it became possible to add new users and send the latest news to all users.
The file was deleted from the numbers of the last sent news.
The get_user_key file has been changed: now the same script records the chat id not only the user's name, but also the last news sent to him.