topics 0.5
The main focus was on getting rid of a lot of list in the code. As a result, I came to the implementation via pandas DataFrame (as it seemed to me it looks more concise).
In addition, the following error has been corrected: "when you select a topic that is in several categories, it is duplicated."  Now each topic is assigned a unique key taken from the data parsing source. So now if the user selects the "others" of the "medicine" section, then he will be offered news only from this section, and not as before all the "others" in a row.
It was decided to get rid of the possibility of selecting all the topics at the same time (I didn't like the way it looked).